+++ client/app/account/IngredientInput.js
@@ -3,7 +3,7 @@
 import React, { useState } from 'react';
 import { CustomTextField, CustomDropdown } from "../components/customComponents.js"
 import { AddCircleOutlineRounded, RemoveCircleOutlineRounded } from '@mui/icons-material';
-import styles from './pantry.module.css'
+import styles from './account.module.css'
 import { MenuItem } from '@mui/material';
 
 const IngredientInput = () => {
