+++ client/app/account/IngredientList.js
@@ -3,7 +3,7 @@
 import React, { useEffect, useState } from 'react';
 import { MenuItem, IconButton, TextField, Button } from '@mui/material';
 import { RemoveCircleOutlineRounded, EditRounded, SaveRounded, CloseRounded } from '@mui/icons-material';
-import styles from './pantry.module.css';
+import styles from './account.module.css';
 import { CustomTextField, CustomDropdown } from "../components/customComponents.js"
 
 const capitalizeFirstLetter = (string) => {
