+++ client/app/account/NutritionInput.js
@@ -1,7 +1,7 @@
 "use client";
 
 import React, { useState, useEffect } from 'react';
-import styles from './pantry.module.css';
+import styles from './account.module.css';
 import { CustomTextField, CustomCircularProgress } from "../components/customComponents.js";
 
 const colorMapping = {
