--- client/app/plan/PlanPage.module.css
@@ -1,62 +1,0 @@
-.planPageWrapper {
-    width: 100%;
-    height: 100%;
-    padding: 2em 8em; /* Add some padding for better layout */
-}
-
-.title {
-    margin-top: 3.5em;
-}
-
-.recipesContainer {
-    display: grid;
-    grid-template-columns: repeat(4, 1fr);
-    grid-column-gap: 5em;
-    grid-row-gap: 3em;
-    margin-top: 2em;
-}
-
-.recipeCard {
-    background-color: rgba(171, 169, 231, 0.2);
-    border-radius: 10px;
-    padding: .8em 1em;
-    height: 24vh;
-    width: 16.4vw;
-    cursor: pointer;
-    text-decoration: none;
-    transition: background-color 0.3s, box-shadow 0.3s; /* Smooth transition */
-}
-
-.recipeCard:hover {
-    background-color: rgba(171, 169, 231, 0.264);
-    box-shadow: 0 4px 10px rgba(65, 65, 82, 0.3);
-}
-
-.recipeTitleWrapper {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    color: #506264;
-}
-
-.recipeTitle {
-    font-weight: 600;
-    margin-top: .3em;
-    font-size: 1em;
-    max-width: 80%;
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-}
-
-.recipeImage {
-    height: 75%;
-    width: 100%;
-    border-radius: 10px;
-    object-fit: cover;
-}
-
-.recipeClock {
-    font-size: 1.4em !important;
-    margin-right: .3em;
-}
