+++ client/app/plan/page.js
@@ -1,106 +1,106 @@
-'use client';
+// 'use client';
 
-import { useEffect, useState } from 'react';
-import { useRouter } from 'next/navigation';
-import Link from 'next/link';
-import Navbar from '../components/navbar';
-import styles from './PlanPage.module.css';
-import AccessTimeIcon from '@mui/icons-material/AccessTime'; // Importing the icon
-import LocalDiningIcon from '@mui/icons-material/LocalDining';
+// import { useEffect, useState } from 'react';
+// import { useRouter } from 'next/navigation';
+// import Link from 'next/link';
+// import Navbar from '../components/navbar';
+// import styles from './PlanPage.module.css';
+// import AccessTimeIcon from '@mui/icons-material/AccessTime'; // Importing the icon
+// import LocalDiningIcon from '@mui/icons-material/LocalDining';
 
-const Plan = () => {
-    const [loading, setLoading] = useState(true);
-    const [recipes, setRecipes] = useState([]);
-    const [error, setError] = useState(null);
-    const router = useRouter();
+// const Plan = () => {
+//     const [loading, setLoading] = useState(true);
+//     const [recipes, setRecipes] = useState([]);
+//     const [error, setError] = useState(null);
+//     const router = useRouter();
 
-    // Fetch user's recipes
-    const fetchUserRecipes = async (token) => {
-        try {
-            const response = await fetch(process.env.NEXT_PUBLIC_SERVER_URL + '/api/plan/user-recipes', {
-                headers: {
-                    Authorization: `Bearer ${token}`,
-                },
-            });
+//     // Fetch user's recipes
+//     const fetchUserRecipes = async (token) => {
+//         try {
+//             const response = await fetch(process.env.NEXT_PUBLIC_SERVER_URL + '/api/plan/user-recipes', {
+//                 headers: {
+//                     Authorization: `Bearer ${token}`,
+//                 },
+//             });
 
-            const data = await response.json();
-            setRecipes(data.recipes);
-        } catch (error) {
-            console.error('Error fetching user recipes:', error);
-            setError('Failed to load your recipes.');
-        } finally {
-            setLoading(false);
-        }
-    };
+//             const data = await response.json();
+//             setRecipes(data.recipes);
+//         } catch (error) {
+//             console.error('Error fetching user recipes:', error);
+//             setError('Failed to load your recipes.');
+//         } finally {
+//             setLoading(false);
+//         }
+//     };
 
-    useEffect(() => {
-        const token = localStorage.getItem('token');
-        const verifyToken = async () => {
-            if (!token) {
-                setLoading(false);
-                router.push('/login');  // Redirect if token is missing
-                return;
-            }
+//     useEffect(() => {
+//         const token = localStorage.getItem('token');
+//         const verifyToken = async () => {
+//             if (!token) {
+//                 setLoading(false);
+//                 router.push('/login');  // Redirect if token is missing
+//                 return;
+//             }
 
-            try {
-                const response = await fetch(`${process.env.NEXT_PUBLIC_SERVER_URL}/api/protected`, {
-                    method: 'GET',
-                    headers: {
-                        'Authorization': `Bearer ${token}`,
-                    },
-                });
+//             try {
+//                 const response = await fetch(`${process.env.NEXT_PUBLIC_SERVER_URL}/api/protected`, {
+//                     method: 'GET',
+//                     headers: {
+//                         'Authorization': `Bearer ${token}`,
+//                     },
+//                 });
 
-                if (response.ok) {
-                    await fetchUserRecipes(token);  // Fetch user recipes if token is valid
-                } else {
-                    router.push('/login'); // Redirect if token is invalid
-                }
-            } catch (error) {
-                console.error('Error verifying token:', error);
-                router.push('/login');  // Redirect on error
-            } finally {
-                setLoading(false);
-            }
-        };
+//                 if (response.ok) {
+//                     await fetchUserRecipes(token);  // Fetch user recipes if token is valid
+//                 } else {
+//                     router.push('/login'); // Redirect if token is invalid
+//                 }
+//             } catch (error) {
+//                 console.error('Error verifying token:', error);
+//                 router.push('/login');  // Redirect on error
+//             } finally {
+//                 setLoading(false);
+//             }
+//         };
 
-        verifyToken();
-    }, [router]);
+//         verifyToken();
+//     }, [router]);
 
-    if (loading) {
-        return <div>Loading...</div>;
-    }
+//     if (loading) {
+//         return <div>Loading...</div>;
+//     }
 
-    if (error) {
-        return <div>Error: {error}</div>;
-    }
+//     if (error) {
+//         return <div>Error: {error}</div>;
+//     }
 
-    return (
-        <div>
-            <Navbar />
-            <div className={styles.planPageWrapper}>
-                <div className={styles.title}>
-                    <h1>Your Plan</h1>
-                </div>
-                <div className={styles.recipesContainer}>
-                    {recipes.length > 0 ? (
-                        recipes.map((recipe) => (
-                            <Link href={`/plan/${recipe.id}`} key={recipe.id} className={styles.recipeCard}>
-                                <img src={recipe.image} alt={recipe.title} className={styles.recipeImage} />
-                                <div className={styles.recipeTitleWrapper}>
-                                    <div className={styles.recipeTitle}>{recipe.title}</div>
-                                    <AccessTimeIcon className={styles.recipeClock} />{recipe.readyInMinutes} min
-                                    <LocalDiningIcon className={styles.recipeClock} />
-                                        {recipe.extendedIngredients?.length || 0} Ingredients
-                                </div>
-                            </Link>
-                        ))
-                    ) : (
-                        <p>No recipes in your plan.</p>
-                    )}
-                </div>
-            </div>
-        </div>
-    );
-};
+//     return (
+//         <div>
+//             <Navbar />
+//             <div className={styles.planPageWrapper}>
+//                 <div className={styles.title}>
+//                     <h1>Your Plan</h1>
+//                 </div>
+//                 <div className={styles.recipesContainer}>
+//                     {recipes?.length > 0 ? (
+//                         recipes.map((recipe) => (
+//                             <Link href={`/plan/${recipe.id}`} key={recipe.id} className={styles.recipeCard}>
+//                                 <img src={recipe.image} alt={recipe.title} className={styles.recipeImage} />
+//                                 <div className={styles.recipeTitleWrapper}>
+//                                     <div className={styles.recipeTitle}>{recipe.title}</div>
+//                                     <AccessTimeIcon className={styles.recipeClock} />{recipe.readyInMinutes} min
+//                                     <LocalDiningIcon className={styles.recipeClock} />
+//                                         {recipe.extendedIngredients?.length || 0} Ingredients
+//                                 </div>
+//                             </Link>
+//                         ))
+//                     ) : (
+//                         <p>No recipes in your plan.</p>
+//                     )}
+//                 </div>
+//             </div>
+//         </div>
+//     );
+// };
 
-export default Plan;
+// export default Plan;
